# Conflicts:
#	src/main/java/com/company/myapp/controller/EmployeeController.java

+++ src/main/java/com/company/myapp/controller/EmployeeController.java
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/emp")
 public class EmployeeController {
 
-    private static final String SUCCESS = "success";
     @Autowired
     private EmployeeService employeeService;
 
@@ -21,7 +20,7 @@ public class EmployeeController {
     private UserService userService;
 
     @GetMapping
-    public String getEmployees(Model model) {
+    public String getEmployees(Model model){
         employeeService.getAll(model);
         employeeService.getUsersToAdd(model);
         employeeService.getUsersToDelete(model);
@@ -30,15 +29,27 @@ public class EmployeeController {
     }
 
     @PostMapping(value = "/add")
-    public String addEmployee(Model model, @RequestParam("id") Long id) {
-        employeeService.addEmployee(model, id);
-        return SUCCESS;
+    public String addEmployee(Model model, @RequestParam(value = "id", required=false) Long id){
+        try {
+            if(id == null) throw new Exception();
+            employeeService.addEmployee(model, id);
+        } catch (Exception e){
+            model.addAttribute("error", "User was not selected");
+            return "error";
+        }
+        return "success";
     }
 
     @PostMapping(value = "/delete")
-    public String deleteEmployee(Model model, @RequestParam("id") Long id) {
-        employeeService.deleteEmployee(model, id);
-        return SUCCESS;
+    public String deleteEmployee(Model model, @RequestParam(value = "id", required=false) Long id){
+        try {
+            if(id == null) throw new Exception();
+            employeeService.deleteEmployee(model, id);
+        } catch (Exception e){
+            model.addAttribute("error", "User was not selected");
+            return "error";
+        }
+        return "success";
     }
 //    @GetMapping(value = "/add")
 //    public String getUsersToAdd(Model model){
